Conflicts:
	ProjetoIntegrador/src/main/java/br/senac/pi4/ProjetoIntegrador/Service/Implement/ServiceCategoriaImplement.java
	ProjetoIntegrador/target/maven-status/maven-compiler-plugin/compile/default-compile/inputFiles.lst
	ProjetoIntegrador/target/maven-status/maven-compiler-plugin/testCompile/default-testCompile/inputFiles.lst

+++ ProjetoIntegrador/src/main/java/br/senac/pi4/ProjetoIntegrador/Service/CategoriaService.java
@@ -17,11 +17,11 @@ public interface CategoriaService extends Serializable {
     
     public List<Categoria> listar();
     
-    public Categoria obter(long id);
+    public Categoria obter(Long id);
     
     public void incluir(Categoria categoria);
     
     public void alterar(Categoria categoria);
     
-    public void remover(Integer codigoCategoria);
+    public void remover(Long codigoCategoria);
 }
